refactor: guardian 폴더 분리, exception 등 코드리뷰 반영
