+++ src/main/java/dbdr/controller/AdminGuardiansController.java
@@ -1,9 +1,12 @@
-package dbdr.domain.guardians;
+package dbdr.controller;
 
+import dbdr.dto.request.GuardiansRequest;
+import dbdr.dto.response.GuardiansResponse;
+import dbdr.service.GuardiansService;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.annotations.Comment;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,19 +19,16 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/v1/admin/guardians")
 @RequiredArgsConstructor
-@Comment("관리자 페이지의 보호자 관리")
 public class AdminGuardiansController {
 
     private final GuardiansService guardiansService;
 
-    @Comment("보호자 전부 조회")
     @GetMapping
     public ResponseEntity<List<GuardiansResponse>> showAllGuardians() {
         List<GuardiansResponse> guardiansResponseList = guardiansService.getAllGuardians();
         return ResponseEntity.ok(guardiansResponseList);
     }
 
-    @Comment("보호자 한 사람 조회")
     @GetMapping("/{guardianId}")
     public ResponseEntity<GuardiansResponse> showOneGuardian(
         @PathVariable("guardianId") Long guardianId) {
@@ -36,15 +36,13 @@ public class AdminGuardiansController {
         return ResponseEntity.ok(guardiansResponse);
     }
 
-    @Comment("보호자 한 사람 생성")
     @PostMapping
     public ResponseEntity<GuardiansResponse> addGuardian(
-        @Valid @RequestBody GuardiansRequest guardiansRequest) throws Exception {
+        @Valid @RequestBody GuardiansRequest guardiansRequest) {
         GuardiansResponse guardiansResponse = guardiansService.addGuardian(guardiansRequest);
-        return ResponseEntity.ok(guardiansResponse);
+        return ResponseEntity.status(HttpStatus.CREATED).body(guardiansResponse);
     }
 
-    @Comment("보호자 정보 수정")
     @PutMapping("/{guardianId}/update")
     public ResponseEntity<GuardiansResponse> updateGuardianAuth(
         @PathVariable("guardianId") Long guardianId,
@@ -54,11 +52,10 @@ public class AdminGuardiansController {
         return ResponseEntity.ok(guardiansResponse);
     }
 
-    @Comment("보호자 한 사람 삭제")
     @PutMapping("/{guardianId}/delete")
     public ResponseEntity<Void> deleteGuardianAuth(
         @PathVariable("guardianId") Long guardianId) {
         guardiansService.deleteGuardianById(guardianId);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
 }
