+++ src/main/java/dbdr/repository/GuardiansRepository.java
@@ -1,5 +1,6 @@
-package dbdr.domain.guardians;
+package dbdr.repository;
 
+import dbdr.domain.Guardians;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GuardiansRepository extends JpaRepository<Guardians, Long> {
