+++ src/main/java/dbdr/service/GuardiansService.java
@@ -1,7 +1,10 @@
-package dbdr.domain.guardians;
+package dbdr.service;
 
+import dbdr.domain.Guardians;
+import dbdr.dto.request.GuardiansRequest;
+import dbdr.dto.response.GuardiansResponse;
+import dbdr.repository.GuardiansRepository;
 import java.util.List;
-import org.hibernate.annotations.Comment;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -13,13 +16,11 @@ public class GuardiansService {
         this.guardiansRepository = guardiansRepository;
     }
 
-    @Comment("보호자 아이디로 조회하기")
     public GuardiansResponse getGuardianById(Long guardianId) {
         Guardians guardian = guardiansRepository.findById(guardianId).orElseThrow();
         return new GuardiansResponse(guardian.getPhone(), guardian.getName(), guardian.isActive());
     }
 
-    @Comment("보호자 아이디로 수정하기")
     public GuardiansResponse updateGuardianById(Long guardianId,
         GuardiansRequest guardiansRequest) {
         Guardians guardian = guardiansRepository.findById(guardianId).orElseThrow();
@@ -28,7 +29,6 @@ public class GuardiansService {
         return new GuardiansResponse(guardiansRequest.phone(), guardiansRequest.name(), guardian.isActive());
     }
 
-    @Comment("보호자 전부 조회하기")
     public List<GuardiansResponse> getAllGuardians() {
         List<Guardians> guardians = guardiansRepository.findAll();
         return guardians.stream()
@@ -36,24 +36,22 @@ public class GuardiansService {
             .toList();
     }
 
-    @Comment("보호자 한 사람 새로 추가하기")
-    public GuardiansResponse addGuardian(GuardiansRequest guardiansRequest) throws Exception {
+    public GuardiansResponse addGuardian(GuardiansRequest guardiansRequest) {
         phoneNumberExists(guardiansRequest.phone());
         Guardians guardian = new Guardians(guardiansRequest.phone(), guardiansRequest.name());
         guardian = guardiansRepository.save(guardian);
         return new GuardiansResponse(guardian.getPhone(), guardian.getName(), guardian.isActive());
     }
 
-    @Comment("보호자 한 사람 삭제하기(DB에는 존재)")
     public void deleteGuardianById(Long guardianId) {
         Guardians guardians = guardiansRepository.findById(guardianId).orElseThrow();
         guardians.deactivate();
         guardiansRepository.save(guardians);
     }
 
-    private void phoneNumberExists(String phone) throws Exception {
+    private void phoneNumberExists(String phone) {
         if(guardiansRepository.existsByPhone(phone)){
-            throw new Exception("존재하는 전화번호입니다.");
+            throw new IllegalArgumentException("존재하는 전화번호입니다.");
         }
     }
 }
